Conflicts:
	src/java/controllers/WallController.java
	web/WEB-INF/jsp/createGame.jsp
	web/WEB-INF/jsp/wall.jsp

+++ src/java/services/MessageService.java
@@ -14,5 +14,5 @@ import dao.NotificationEntity;
  */
 public interface MessageService {
     public Boolean addNotification(NotificationEntity n);
-    public Boolean addMessage(String contenu, String login, Long idQuestion);
+    public Boolean addMessage(String contenu, String login, Integer idQuestion);
 }
